refactor: move WebConfig to config/ and UploadResponse to model/dto/
